Renamed `events.rs` file to `event.rs`
Suggested by clippy
