+++ wolf_engine_core/src/events/mod.rs
@@ -5,5 +5,5 @@
 
 pub use generic_event_queue::*;
 
-mod events;
-pub use events::*;
+mod event;
+pub use event::*;
